#FEM-1024 - Changed to Int32 from Float in currentTIme. (#45)

+++ Plugins/Phoenix/PhoenixAnalyticsPlugin.swift
@@ -68,7 +68,7 @@ public class PhoenixAnalyticsPlugin: PKPlugin, KalturaPluginManagerDelegate {
                                                                        partnerId: parterId,
                                                                        ks: ks,
                                                                        eventType: action.rawValue.uppercased(),
-                                                                       currentTime: Float(self.player.currentTime),
+                                                                       currentTime: Int32(self.player.currentTime),
                                                                        assetId: self.mediaEntry.id,
                                                                        fileId: fileId) {
             builder.set { (response: Response) in
