+++ Plugins/Phoenix/TVPAPIAnalyticsPlugin.swift
@@ -67,7 +67,7 @@ public class TVPAPIAnalyticsPlugin: PKPlugin, KalturaPluginManagerDelegate {
         if let builder: RequestBuilder = MediaMarkService.sendTVPAPIEVent(baseURL: baseUrl,
                                                                                  initObj: initObj,
                                                                                  eventType: action.rawValue,
-                                                                                 currentTime: Float(self.player.currentTime),
+                                                                                 currentTime: Int32(self.player.currentTime),
                                                                                  assetId: self.mediaEntry.id,
                                                                                  fileId: fileId) {
             builder.set { (response: Response) in
